crearOportunidadDocente agregado a f1

+++ src/main/java/com/vias/uc/backend/repository/OportunidadRepository.java
@@ -3,7 +3,4 @@ package com.vias.uc.backend.repository;
 import com.vias.uc.backend.model.Oportunidad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OportunidadRepository extends JpaRepository<Oportunidad, Long> {}
-
-
-
+public interface OportunidadRepository extends JpaRepository<Oportunidad, Integer> {}
